Got compiling without warnings on vlg cluster. Must now use HAVE_OPT++ for the gradient-based optimization (not yet working on vlg cluster).

+++ MeshIO.hh
@@ -194,7 +194,7 @@ namespace MeshIO {
             typedef IOElement Element;
 
             void getElementInfo(MeshType meshType, int &elementType,
-                                int &numCorners) {
+                                size_t &numCorners) {
                 switch (meshType) {
                     case MESH_TRI:
                         elementType = 2;
